[FEATURE] Add generic ConstantsToBackedEnumValueRector
Relates: #3698
