+++ src/Contract/NoChangelogRequiredInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Ssch\TYPO3Rector\Contract;
+
+interface NoChangelogRequiredInterface
+{
+}
