+++ tests/Rector/General/Renaming/ConstantsToBackedEnumValueRector/Source/MyOtherFileType.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Ssch\TYPO3Rector\Tests\Rector\v13\v0\MigrateFileTypeConstantsToFileTypeEnumRector\Source;
+namespace Ssch\TYPO3Rector\Tests\Rector\General\Renaming\ConstantsToBackedEnumValueRector\Source;
 
 final class MyOtherFileType
 {
