+++ tests/Rector/General/Renaming/ConstantsToBackedEnumValueRector/config/configured_rule.php
@@ -4,11 +4,12 @@ declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
 use Rector\Renaming\ValueObject\RenameClassAndConstFetch;
+use Rector\ValueObject\PhpVersion;
 use Ssch\TYPO3Rector\General\Renaming\ConstantsToBackedEnumValueRector;
 
 return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->import(__DIR__ . '/../config.php');
-
+    $rectorConfig->import(__DIR__ . '/../../../../../../config/config.php');
+    $rectorConfig->phpVersion(PhpVersion::PHP_81);
     $rectorConfig->ruleWithConfiguration(ConstantsToBackedEnumValueRector::class, [
         new RenameClassAndConstFetch(
             'TYPO3\CMS\Core\Resource\AbstractFile',
@@ -46,37 +47,5 @@ return static function (RectorConfig $rectorConfig): void {
             'TYPO3\CMS\Core\Resource\FileType',
             'APPLICATION'
         ),
-
-        new RenameClassAndConstFetch(
-            'TYPO3\CMS\Core\Imaging\Icon',
-            'SIZE_DEFAULT',
-            'TYPO3\CMS\Core\Imaging\IconSize',
-            'DEFAULT'
-        ),
-        new RenameClassAndConstFetch(
-            'TYPO3\CMS\Core\Imaging\Icon',
-            'SIZE_SMALL',
-            'TYPO3\CMS\Core\Imaging\IconSize',
-            'SMALL'
-        ),
-        new RenameClassAndConstFetch(
-            'TYPO3\CMS\Core\Imaging\Icon',
-            'SIZE_MEDIUM',
-            'TYPO3\CMS\Core\Imaging\IconSize',
-            'MEDIUM'
-        ),
-        new RenameClassAndConstFetch(
-            'TYPO3\CMS\Core\Imaging\Icon',
-            'SIZE_LARGE',
-            'TYPO3\CMS\Core\Imaging\IconSize',
-            'LARGE'
-        ),
-        new RenameClassAndConstFetch(
-            'TYPO3\CMS\Core\Imaging\Icon',
-            'SIZE_MEGA',
-            'TYPO3\CMS\Core\Imaging\IconSize',
-            'MEGA'
-        ),
-
     ]);
 };
